Page Components created, HW3 done
